requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
tensorflow==2.19.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0